Extract DirectoryContents from utils:files
Bug: b/540479740
