+++ base/cvd/cuttlefish/common/libs/utils/BUILD.bazel
@@ -115,6 +115,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:in_sandbox",
         "//cuttlefish/common/libs/utils:users",
         "//cuttlefish/files:copy",
+        "//cuttlefish/files:directory_contents",
         "//cuttlefish/files:file_device_id",
         "//cuttlefish/files:file_exists",
         "//cuttlefish/files:link_or_copy",
