+++ base/cvd/cuttlefish/common/libs/utils/files.h
@@ -42,7 +42,6 @@ Result<void> LinkOrCopyDirectoryContentsRecursively(
 Result<void> MoveDirectoryContents(const std::string& source,
                                    const std::string& destination);
 bool FileHasContent(const std::string& path);
-Result<std::vector<std::string>> DirectoryContents(const std::string& path);
 Result<std::vector<std::string>> DirectoryContentsPaths(
     const std::string& path);
 bool DirectoryExists(const std::string& path, bool follow_symlinks = true);
