+++ base/cvd/cuttlefish/files/BUILD.bazel
@@ -30,6 +30,16 @@ cf_cc_library(
 )
 
 cf_cc_library(
+    name = "directory_contents",
+    srcs = ["directory_contents.cc"],
+    hdrs = ["directory_contents.h"],
+    deps = [
+        "//cuttlefish/result:expect",
+        "//cuttlefish/result:result_type",
+    ],
+)
+
+cf_cc_library(
     name = "file_device_id",
     srcs = ["file_device_id.cc"],
     hdrs = ["file_device_id.h"],
