+++ base/cvd/cuttlefish/host/commands/assemble_cvd/android_build/BUILD.bazel
@@ -56,9 +56,8 @@ cf_cc_library(
     name = "android_product_dir",
     srcs = ["android_product_dir.cc"],
     hdrs = ["android_product_dir.h"],
-    depend_on_what_you_use_enabled = False,
     deps = [
-        "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:directory_contents",
         "//cuttlefish/files:file_exists",
         "//cuttlefish/host/commands/assemble_cvd/android_build",
         "//cuttlefish/pretty:result",
@@ -66,7 +65,6 @@ cf_cc_library(
         "//cuttlefish/pretty:struct",
         "//cuttlefish/result",
         "@abseil-cpp//absl/strings",
-        "@abseil-cpp//absl/strings:str_format",
     ],
 )
 
@@ -138,7 +136,7 @@ cf_cc_library(
     srcs = ["find_build_archive.cc"],
     hdrs = ["find_build_archive.h"],
     deps = [
-        "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:directory_contents",
         "//cuttlefish/host/libs/config:build_archive",
         "//cuttlefish/host/libs/config:fetcher_config",
         "//cuttlefish/host/libs/config:file_source",
