+++ base/cvd/cuttlefish/host/commands/cvd/cli/BUILD.bazel
@@ -72,6 +72,7 @@ cf_cc_library(
     hdrs = ["log_files.h"],
     deps = [
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:directory_contents",
         "//cuttlefish/host/commands/cvd/utils",
         "//cuttlefish/result",
         "@abseil-cpp//absl/log",
@@ -84,6 +85,7 @@ cf_cc_test(
     srcs = ["log_files_test.cpp"],
     deps = [
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:directory_contents",
         "//cuttlefish/files:file_exists",
         "//cuttlefish/files:recursively_remove_directory",
         "//cuttlefish/host/commands/cvd/cli:log_files",
