+++ base/cvd/cuttlefish/host/commands/cvd/cli/log_files.cpp
@@ -28,6 +28,7 @@
 #include "absl/strings/match.h"
 
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/files/directory_contents.h"
 #include "cuttlefish/host/commands/cvd/utils/common.h"
 
 namespace cuttlefish {
