+++ base/cvd/cuttlefish/host/libs/config/host_tools_version.cpp
@@ -28,6 +28,7 @@
 #include <zlib.h>
 
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/files/directory_contents.h"
 #include "cuttlefish/host/libs/config/config_utils.h"
 
 namespace cuttlefish {
