+++ base/cvd/cuttlefish/process/BUILD.bazel
@@ -68,6 +68,7 @@ cf_cc_library(
     deps = [
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:directory_contents",
         "//cuttlefish/posix:readlink",
         "//cuttlefish/result",
         "//libbase",
