+++ base/cvd/cuttlefish/process/proc_file_utils.cc
@@ -40,6 +40,7 @@
 #include "cuttlefish/common/libs/fs/shared_buf.h"
 #include "cuttlefish/common/libs/fs/shared_fd.h"
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/files/directory_contents.h"
 #include "cuttlefish/posix/readlink.h"
 #include "cuttlefish/result/result.h"
